Add files via upload
This program asks you to input a number. Then it will tell you if the given number is a palindrome, if not then it will tell you the next number that will be a palindrome.
(Palindrome: A word which is same no matter whether it is read from right side or left. E.g.: Mom, Malayalam, Pup etc. )